pyproject.toml
[project]
name = "flask-crud"
version = "0.1.0"
description = ""
authors = [
	{ name = "Anirudh Ramakrishnan", email = "38531113+anirudh-r5@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"flask (>=3.1.0,<4.0.0)",
	"python-dotenv (>=1.0.1,<2.0.0)",
	"watchdog (>=6.0.0,<7.0.0)",
	"psycopg[binary] (>=3.2.6,<4.0.0)",
	"pyjwt (>=2.10.1,<3.0.0)",
	"bcrypt (>=4.3.0,<5.0.0)",
]

[tool.poetry]
packages = [{ include = "flask_crud", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
